Added CollatzSequencer
Performs collatz conjecture.
Default: Starts with "start" number then follows collatz sequence to 1 and returns list for sequence
Return Highest: Starts with "start" number then follows collatz sequence to 1 and returns the highest number in the sequence

All functions that are exclusively for integers now convert the input to an integer for ease of use.