Added LCM and Dependencies
A LCM function has been added using FindFactors which has also been added which is a tweaked version of FindNoOfFactors which is added as well. LowestCommonMultipleWIP had some ... bugs. This should be faster for large inputs as it will use the prime factors of the inputs but this is not finished yet so shouldn't be used unless it is to your program.