Added FindPrimeFactors and removed IsPrime
The temporary file used for testing has been added as well as the function to APE Functions. IsPrime3 has now replaced IsPrime (version 2) and is called IsPrime. "IsPrime3" is still in the library temporarily to let programs transition to IsPrime. More work was done on LCM WIP which will use the FindPrimeFactors functions.