Non-Specific Functions Moved To Other Libraries
getMin/MaxValue have been moved to the StandardFunctions library to allow use in other programs. EquationsArray objects can now be created instead of using a standard array to store the data. This allows for encapsulation of all the relavent functions and allows it to be used by other programs without importing the graph plotting library. This is still being developed and has broken functionality temporarily.