Graph can be panned
By changing the way the screen is related to the world and adding in a buch of screen-to-world movement functions, the arrow keys can be used to essentially move the grpah around to see other parts of the plotted area. The graph does seem to get squished currently and you can't hold a key down.